FIX #9243: l10n_ve_accountant, l10n_ve_payment_extension
Error en el flujo de retenciones del IVA al eliminar una línea y al seleccionar las facturas disponibles

Se corrigio la creacion de los pagos en las retenciones, el mismo se crea al confirmar la retencion y no al momento de crear las lineas de retencion
Se corrio eliminacion de las lineas de retenciones al crear campos para mantener los ids de facturas validas para la retencion del cliente y pasarlos como parametros de dominio
Se corrio el dominio de las facturas en las vistas
Se agrego validacion en para que las lineas de retencion se bloqueen si no existe un partner elegido

Tarea (Link):https://binaural.odoo.com/web#id=9243&cids=2&menu_id=293&action=393&active_id=58&model=helpdesk.ticket&view_type=form

Tarea de proyecto []
Ticket de soporte [x]

+++ l10n_ve_payment_extension/models/account_retention_line.py
@@ -7,8 +7,11 @@ _logger = logging.getLogger(__name__)
 class AccountRetentionLine(models.Model):
     _name = "account.retention.line"
     _description = "Retention Line"
+    _rec_name = 'name'
 
     check_company = True
+
+    
 
     name = fields.Char(
         string="Description", required=True, compute="_compute_name", store=True, readonly=False
@@ -55,7 +58,7 @@ class AccountRetentionLine(models.Model):
     iva_amount = fields.Float(string="IVA", digits=(16, 2))
 
     retention_amount = fields.Float(
-        digits="Tasa", compute="_compute_retention_amount", store=True, readonly=False
+        digits="Tasa", compute="_compute_retention_amount_compute_name", store=True, readonly=False
     )
     foreign_retention_amount = fields.Float(
         digits="Tasa", compute="_compute_retention_amount", store=True, readonly=False
@@ -157,7 +160,8 @@ class AccountRetentionLine(models.Model):
 
     def unlink(self):
         for record in self:
-            record.payment_id.unlink()
+            if record.payment_id:
+                record.payment_id.unlink()
         return super().unlink()
 
     @api.onchange("payment_concept_id")
